[chore/#3] 패키지명 변경
