+++ composeApp/src/androidMain/kotlin/com/nexters/ace/MainActivity.kt
@@ -1,4 +1,4 @@
-package org.nexters.ace
+package com.nexters.ace
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
