+++ composeApp/src/androidMain/kotlin/com/nexters/ace/Platform.android.kt
@@ -1,4 +1,4 @@
-package org.nexters.ace
+package com.nexters.ace
 
 import android.os.Build
