+++ composeApp/src/commonMain/kotlin/com/nexters/ace/Platform.kt
@@ -1,0 +1,7 @@
+package com.nexters.ace
+
+interface Platform {
+    val name: String
+}
+
+expect fun getPlatform(): Platform
