--- composeApp/src/commonMain/kotlin/org/nexters/ace/Platform.kt
@@ -1,7 +1,0 @@
-package org.nexters.ace
-
-interface Platform {
-    val name: String
-}
-
-expect fun getPlatform(): Platform
