+++ composeApp/src/iosMain/kotlin/com/nexters/ace/MainViewController.kt
@@ -1,0 +1,5 @@
+package com.nexters.ace
+
+import androidx.compose.ui.window.ComposeUIViewController
+
+fun MainViewController() = ComposeUIViewController { App() }
