+++ composeApp/src/iosMain/kotlin/com/nexters/ace/Platform.ios.kt
@@ -1,0 +1,10 @@
+package com.nexters.ace
+
+import platform.UIKit.UIDevice
+
+class IOSPlatform : Platform {
+    override val name: String =
+        UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
+}
+
+actual fun getPlatform(): Platform = IOSPlatform()
