--- composeApp/src/iosMain/kotlin/org/nexters/ace/MainViewController.kt
@@ -1,5 +1,0 @@
-package org.nexters.ace
-
-import androidx.compose.ui.window.ComposeUIViewController
-
-fun MainViewController() = ComposeUIViewController { App() }
