--- composeApp/src/iosMain/kotlin/org/nexters/ace/Platform.ios.kt
@@ -1,9 +1,0 @@
-package org.nexters.ace
-
-import platform.UIKit.UIDevice
-
-class IOSPlatform: Platform {
-    override val name: String = UIDevice.currentDevice.systemName() + " " + UIDevice.currentDevice.systemVersion
-}
-
-actual fun getPlatform(): Platform = IOSPlatform()
